LogIn Portal
Built a very simplified desktop app, with the use of visual studio and xcode.

Programming language used C#.

UI login window created to take the following inputs from user:

-Username
-Password
-Remember me (checkbox)
-Login (action button)

User input is printed in Application Output console.

This done by declaring variables and assigning them to values.

Then storing in one variable which is a string and then printed this information to the console.